# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LEGENDLEO Gamehub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Game {
  id: number;
  title: string;
  description: string;
  category: string;
  imageUrl: string;
  rating: number;
  platform: string;
  size: string;
  sizeMb?: number;
  version: string;
  downloadUrl: string;
  featured: boolean;
  popular: boolean;
  modFeatures: string[];
  downloadCount: number;
  createdAt: string;
}

export interface GameTopItem {
  id: number;
  title: string;
  downloadCount: number;
  category: string;
  imageUrl: string;
}

export interface CategoryStat {
  name: string;
  count: number;
  totalDownloads: number;
}

export interface GameStats {
  totalGames: number;
  totalCategories: number;
  totalDownloads: number;
  featuredCount: number;
  popularCount: number;
  topGames: GameTopItem[];
  categoryBreakdown: CategoryStat[];
}

export interface Category {
  name: string;
  count: number;
}

export interface App {
  id: number;
  title: string;
  description: string;
  category: string;
  developer: string;
  imageUrl: string;
  rating: number;
  platform: string;
  size: string;
  sizeMb?: number;
  version: string;
  downloadUrl: string;
  featured: boolean;
  popular: boolean;
  features: string[];
  downloadCount: number;
  createdAt: string;
}

export interface CreateGameInput {
  title: string;
  description: string;
  category: string;
  imageUrl: string;
  rating?: number;
  platform?: string;
  size?: string;
  sizeMb?: number;
  version?: string;
  downloadUrl?: string;
  featured?: boolean;
  popular?: boolean;
  modFeatures?: string[];
}

export interface CreateAppInput {
  title: string;
  description: string;
  category: string;
  developer?: string;
  imageUrl: string;
  rating?: number;
  platform?: string;
  size?: string;
  sizeMb?: number;
  version?: string;
  downloadUrl?: string;
  featured?: boolean;
  popular?: boolean;
  features?: string[];
}

export interface ErrorResponse {
  error: string;
}

export type ListGamesParams = {
category?: string;
search?: string;
featured?: boolean;
popular?: boolean;
minRating?: number;
sort?: ListGamesSort;
};

export type ListGamesSort = typeof ListGamesSort[keyof typeof ListGamesSort];


export const ListGamesSort = {
  newest: 'newest',
  popular: 'popular',
  rating: 'rating',
  alphabetical: 'alphabetical',
} as const;

export type ListAppsParams = {
category?: string;
search?: string;
featured?: boolean;
popular?: boolean;
minRating?: number;
sort?: ListAppsSort;
};

export type ListAppsSort = typeof ListAppsSort[keyof typeof ListAppsSort];


export const ListAppsSort = {
  newest: 'newest',
  popular: 'popular',
  rating: 'rating',
  alphabetical: 'alphabetical',
} as const;