# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LEGENDLEO Gamehub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: games
    description: Game catalog operations
  - name: categories
    description: Category operations
  - name: apps
    description: App catalog operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all games with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: popular
          in: query
          schema:
            type: boolean
        - name: minRating
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, popular, rating, alphabetical]
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameInput"
      responses:
        "201":
          description: Game created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"

  /games/stats:
    get:
      operationId: getGameStats
      tags: [games]
      summary: Dashboard stats - total games, categories, downloads, top games
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStats"

  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a single game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateGame
      tags: [games]
      summary: Update an existing game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameInput"
      responses:
        "200":
          description: Game updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/{id}/download:
    post:
      operationId: trackDownload
      tags: [games]
      summary: Increment download count for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"

  /apps:
    get:
      operationId: listApps
      tags: [apps]
      summary: List all apps with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: popular
          in: query
          schema:
            type: boolean
        - name: minRating
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, popular, rating, alphabetical]
      responses:
        "200":
          description: List of apps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"
    post:
      operationId: createApp
      tags: [apps]
      summary: Create a new app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppInput"
      responses:
        "201":
          description: App created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"

  /apps/{id}:
    get:
      operationId: getApp
      tags: [apps]
      summary: Get a single app by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: App detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateApp
      tags: [apps]
      summary: Update an existing app
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppInput"
      responses:
        "200":
          description: App updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /apps/{id}/download:
    post:
      operationId: trackAppDownload
      tags: [apps]
      summary: Increment download count for an app
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"

  /app-categories:
    get:
      operationId: listAppCategories
      tags: [apps]
      summary: List all available app categories with counts
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all available game categories with counts
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Game:
      type: object
      required:
        - id
        - title
        - description
        - category
        - imageUrl
        - rating
        - platform
        - size
        - version
        - downloadUrl
        - featured
        - popular
        - modFeatures
        - downloadCount
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        platform:
          type: string
        size:
          type: string
        sizeMb:
          type: integer
        version:
          type: string
        downloadUrl:
          type: string
        featured:
          type: boolean
        popular:
          type: boolean
        modFeatures:
          type: array
          items:
            type: string
        downloadCount:
          type: integer
        createdAt:
          type: string

    GameStats:
      type: object
      required:
        - totalGames
        - totalCategories
        - totalDownloads
        - featuredCount
        - popularCount
        - topGames
        - categoryBreakdown
      properties:
        totalGames:
          type: integer
        totalCategories:
          type: integer
        totalDownloads:
          type: integer
        featuredCount:
          type: integer
        popularCount:
          type: integer
        topGames:
          type: array
          items:
            $ref: "#/components/schemas/GameTopItem"
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryStat"

    GameTopItem:
      type: object
      required:
        - id
        - title
        - downloadCount
        - category
        - imageUrl
      properties:
        id:
          type: integer
        title:
          type: string
        downloadCount:
          type: integer
        category:
          type: string
        imageUrl:
          type: string

    CategoryStat:
      type: object
      required:
        - name
        - count
        - totalDownloads
      properties:
        name:
          type: string
        count:
          type: integer
        totalDownloads:
          type: integer

    Category:
      type: object
      required:
        - name
        - count
      properties:
        name:
          type: string
        count:
          type: integer

    App:
      type: object
      required:
        - id
        - title
        - description
        - category
        - developer
        - imageUrl
        - rating
        - platform
        - size
        - version
        - downloadUrl
        - featured
        - popular
        - features
        - downloadCount
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        developer:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        platform:
          type: string
        size:
          type: string
        sizeMb:
          type: integer
        version:
          type: string
        downloadUrl:
          type: string
        featured:
          type: boolean
        popular:
          type: boolean
        features:
          type: array
          items:
            type: string
        downloadCount:
          type: integer
        createdAt:
          type: string

    CreateGameInput:
      type: object
      required:
        - title
        - description
        - category
        - imageUrl
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        platform:
          type: string
        size:
          type: string
        sizeMb:
          type: integer
        version:
          type: string
        downloadUrl:
          type: string
        featured:
          type: boolean
        popular:
          type: boolean
        modFeatures:
          type: array
          items:
            type: string

    CreateAppInput:
      type: object
      required:
        - title
        - description
        - category
        - imageUrl
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        developer:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        platform:
          type: string
        size:
          type: string
        sizeMb:
          type: integer
        version:
          type: string
        downloadUrl:
          type: string
        featured:
          type: boolean
        popular:
          type: boolean
        features:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string