# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LEGENDLEO Gamehub API
 * OpenAPI spec version: 0.1.0
 */

export interface App {
  id: number;
  title: string;
  description: string;
  category: string;
  developer: string;
  imageUrl: string;
  rating: number;
  platform: string;
  size: string;
  sizeMb?: number;
  version: string;
  downloadUrl: string;
  featured: boolean;
  popular: boolean;
  features: string[];
  downloadCount: number;
  createdAt: string;
}